Refactor LC36 to match new pkg structure
